+++ pytest_girder/pytest_girder/plugin.py
@@ -1,7 +1,10 @@
 from .fixtures import server, db, admin, user
 
+
 def pytest_addoption(parser):
     group = parser.getgroup('girder')
+    group.addoption('--mock-db', action='store_true', default=False,
+                    help='Whether or not to mock the database using mongomock.')
     group.addoption('--mongo-uri', action='store', default='mongodb://localhost:27017',
                     help=('The base URI to the MongoDB instance to use for database connections, '
                           'default is mongodb://localhost:27017'))